feat: add web ui and clarify code sections
