+++ .env.example
@@ -1,2 +1,2 @@
-APRFI_API_KEY=[FILL IN YOUR API KEY HERE]
-API_KEY=[FILL IN YOUR API KEY HERE]
+APRFI_API_KEY=
+API_KEY=
