+++ Dockerfile
@@ -30,4 +30,4 @@ HEALTHCHECK --interval=30s --timeout=10s --start-period=30s --retries=3 \
     CMD curl -f http://localhost:8080/health || exit 1
 
 # Run the application
-CMD ["uvicorn", "hamops.app:app", "--host", "0.0.0.0", "--port", "8080"]
+CMD ["uvicorn", "hamops.main:app", "--host", "0.0.0.0", "--port", "8080"]
