+++ hamops/middleware/__init__.py
@@ -1,0 +1,5 @@
+"""Middleware exports."""
+
+from .logging import RequestLogMiddleware
+
+__all__ = ["RequestLogMiddleware"]
