+++ hamops/models/__init__.py
@@ -1,0 +1,5 @@
+"""Model exports."""
+
+from .callsign import CallsignRecord
+
+__all__ = ["CallsignRecord"]
